+++ src/styles/inputField.scss
@@ -7,23 +7,23 @@
   }
   
   .input-field label {
-    font-size: 14px;
+    font-size: 18px;
     color: #333;
-    margin-bottom: 5px;
+    margin-bottom: 1px;
     font-weight: bold;
   }
   
   .input-field input {
     padding: 10px;
     font-size: 16px;
-    border: 1px solid #ccc;
+    border: 2px solid #007BFF;
     border-radius: 5px;
     outline: none;
     transition: border-color 0.3s ease-in-out;
   }
   
   .input-field input:focus {
-    border-color: #007BFF; /* Example focus color */
+    border-color: #add4ff; 
   }
   
   .input-field input::placeholder {
